feature: add page to view more in depth details for child
# Conflicts:
#	app/src/main/res/layout/fragment_controller_logs.xml
#	app/src/main/res/layout/fragment_rescue_logs.xml

+++ app/src/main/java/com/example/smartair/RescueLogs.java
@@ -1,11 +1,13 @@
 package com.example.smartair;
 
 public class RescueLogs {
-    public int dose;
-    public int preBreathRating;
-    public int postBreathRating;
-    public int postStatus;
+
+    public long dose;
+    public long preBreathRating;
+    public long postBreathRating;
+    public long postStatus;
     public long timestamp;
 
     public RescueLogs() {}
+
 }
